Fix: CommonBaseReservationModal 수정
